paribus_backend/requirements.txt
Django==4.1.4
sqlparse==0.4.3
psycopg2-binary==2.9.5
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
requests==2.28.1
